Move consolidated scan tests into their own TestGroups
