+++ packages/functional-tests/src/test-groups/consolidated-scan-reports-test-group.ts
@@ -1,0 +1,33 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+import { expect } from 'chai';
+import { ScanReport, ScanRunResultResponse } from 'service-library';
+import { TestEnvironment } from '../common-types';
+import { test } from '../test-decorator';
+import { FunctionalTestGroup } from './functional-test-group';
+
+/* eslint-disable @typescript-eslint/no-unused-expressions */
+
+export class ConsolidatedScanReportsTestGroup extends FunctionalTestGroup {
+    @test(TestEnvironment.all)
+    public async testReportCount(): Promise<void> {
+        const response = await this.a11yServiceClient.getScanStatus(this.testContextData.scanId);
+
+        expect((<ScanRunResultResponse>response.body).reports, 'Expected three reports to be returned').to.have.lengthOf(3);
+    }
+
+    @test(TestEnvironment.all)
+    public async testGetReports(): Promise<void> {
+        const response = await this.a11yServiceClient.getScanStatus(this.testContextData.scanId);
+        const reportsInfo = (<ScanRunResultResponse>response.body).reports;
+
+        await Promise.all(
+            reportsInfo.map(async (reportData: ScanReport) => {
+                const reportResponse = await this.a11yServiceClient.getScanReport(this.testContextData.scanId, reportData.reportId);
+
+                this.ensureResponseSuccessStatusCode(response);
+                expect(reportResponse.body, 'Get Scan Report API should return response with defined body').to.not.be.undefined;
+            }),
+        );
+    }
+}
